+++ tests/2xx/220_set_help_text.sh
@@ -3,3 +3,6 @@
 . common.sh
 
 assert_func set_help_text $TEST_OK $TEST_EMPTY "bla"
+assert_func set_help_text $TEST_OK $TEST_EMPTY <<EOF
+blabla
+EOF