+++ tests/2xx/234_add_prereq.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+. common.sh
+
+assert_func add_prereq $TEST_OK $TEST_EMPTY "asfsa\n"
+assert_func add_prereq $TEST_OK $TEST_EMPTY "bash"
+assert_func add_prereq $TEST_FAIL $TEST_EMPTY